separado arquivo de classe Vector3
